Add project initial files
client.py - a client initial exemple
connection.py - an initial class model to encapsulate the connection and synchronize.

server-program - an initial application module for each server computer.
server-program/application.py - an initial server main application.
server-program/server.py - an initial class for encapsulate the server state.
server-program/storage.py - an initial class for encapsulate the shared storage.
server-program/token.py - an initial enum class to describe each command that a server can send to another in the form of a message.

server-program/async - an initial module for managing threads and works
server-program/async/concurrentBag.py - an initial class that implements the producer-consumer pattern
server-program/async/task.py - an initial class that represents the result of a task.
server-program/async/taskManager.py - an initial class that execute functions in async mode using the owner's threads.
server-program/async/taskThread.py - an auxiliary class used by a TaskManager that encapsulate a thread for execute WorkItems.
server-program/async/workException.py - an exception raised when there is a problem when executing a WorkItem in a TaskThread.
server-program/async/workItem.py - an initial class that encapsulates a function method with its arguments and a Task object that represents the result of the encapsulated function when executed.

server-program/async/example.py - an example script using the async module.